refactor: (#75) infrastructure module error package
