+++ simtong-infrastructure/src/main/kotlin/team/comit/simtong/global/error/GlobalErrorHandler.kt
@@ -5,6 +5,7 @@ import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestControllerAdvice
+import team.comit.simtong.global.error.dto.ErrorResponse
 
 /**
  *
