+++ simtong-infrastructure/src/main/kotlin/team/comit/simtong/global/error/WebErrorHandler.kt
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestControllerAdvice
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException
+import team.comit.simtong.global.error.dto.ErrorResponse
 import javax.validation.ConstraintViolationException
 
 /**
